refactor: urlGeneratorForm에서 사용하는 타입을 별도의 type 파일로 분리
